torch_interface - support build.sh custom and build.sh ops_transformer

+++ attention/prompt_flash_attention/torch_interface/build.sh
@@ -10,7 +10,18 @@
 #
 # This script builds the operator and installs a python torch extension package 'torch_pfa'
 
+# Check for exactly one argument
+if [ $# -ne 1 ]; then
+    echo "Error: Exactly one argument required" >&2
+    echo "Usage: $0 <PFA_LIB_value>" >&2
+    echo "  where PFA_LIB_value is either 'custom' or 'ops_transformer'" >&2
+    exit 1
+fi
+
 # build torch extension
 rm -rf build torch_pfa.egg-info
 pip uninstall -y torch_pfa
-pip install . --no-build-isolation
+
+# set PFA_LIB option to be "custom" if you compiled ops-transformer kernlee using "./build.sh --make_clean -j96 --pkg --soc=ascend910b --ops=prompt_flash_attention"
+# set PFA_LIB option to be "ops_transformer" if you compiled ops-transformer kernlee using "./build.sh --make_clean -j96 --pkg --soc=ascend910b"
+PFA_LIB=$1 pip install . --no-build-isolation
